requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.32
python-dotenv>=1.0
pydantic>=2.9

langchain>=0.3
langchain-groq>=0.2
langchain-huggingface>=0.1
langchain-community>=0.3
langchain-chroma>=0.1
langchain-text-splitters>=0.3
langgraph>=0.2
langgraph-checkpoint-sqlite>=2.0
langsmith>=0.1
sentence-transformers>=3.0

chromadb>=0.5
google-api-python-client>=2.140
youtube-transcript-api>=0.6
sse-starlette>=2.1
